requirements.txt
gradio==4.26.0
spaces==0.28.0
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
transformers==4.40.0
qwen-vl-utils==0.0.8
pillow==10.2.0
accelerate==0.27.0
safetensors==0.4.0
huggingface-hub==0.20.1
peft==0.7.1
numpy==1.24.3
